refactor(persistence): move sparse-tokenize into persistence/embeddings
Inverts a persistence/ -> memory/ back-import. The sparse-vector
tokenization primitives (tokenize/tokenizeStemmed/tokenHash/SPARSE_VOCAB_SIZE)
are shared by the TF-only encoder in persistence/embeddings/embedding-backend.ts
and memory's BM25 encoder in v2/sparse-bm25.ts, so they belong in the lowest
layer both reach. embedding-backend.ts stops importing the memory plugin;
its PERSISTENCE_TO_MEMORY_ALLOWLIST entry is removed. Behavior byte-identical.
