+++ assistant/src/__tests__/plugin-import-boundary-guard.test.ts
@@ -123,6 +123,7 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../../persistence/embeddings/embedding-types.js",
     "../../../../persistence/embeddings/qdrant-circuit-breaker.js",
     "../../../../persistence/embeddings/qdrant-client.js",
+    "../../../../persistence/embeddings/sparse-tokenize.js",
     "../../../../persistence/job-utils.js",
     "../../../../persistence/jobs-store.js",
     "../../../../persistence/message-content.js",
@@ -204,7 +205,6 @@ const BASELINE: Record<string, readonly string[]> = {
     "node:fs/promises",
     "node:os",
     "node:path",
-    "stemmer",
     "uuid",
     "yaml",
     "zod",
