+++ assistant/src/persistence/embeddings/__tests__/sparse-tokenize.test.ts
