+++ assistant/src/persistence/embeddings/sparse-tokenize.ts
@@ -3,7 +3,8 @@
 // ---------------------------------------------------------------------------
 //
 // Shared by both the legacy TF-only encoder in `embedding-backend.ts`
-// (`generateSparseEmbedding`) and the BM25 encoder in `v2/sparse-bm25.ts`.
+// (`generateSparseEmbedding`) and the memory BM25 encoder in
+// `plugins/defaults/memory/v2/sparse-bm25.ts`.
 //
 // Lives in its own module so consumers of the BM25 encoder don't transitively
 // depend on `embedding-backend.ts` for these primitives — that matters
@@ -24,9 +25,10 @@ export function tokenize(text: string): string[] {
 /**
  * Tokenize and apply Porter stemming so morphological variants collapse to a
  * shared bucket (e.g. `running`/`runs`/`ran` → `run`, `supplements` →
- * `supplement`). Used only by the BM25 sparse channel in
- * `v2/sparse-bm25.ts`; both the document-side and query-side encoders call
- * this so doc and query tokens land in the same hash buckets.
+ * `supplement`). Used only by the memory BM25 sparse channel in
+ * `plugins/defaults/memory/v2/sparse-bm25.ts`; both the document-side and
+ * query-side encoders call this so doc and query tokens land in the same
+ * hash buckets.
  *
  * Other callers (workspace context-search, the legacy TF-only
  * `generateSparseEmbedding`) intentionally keep the non-stemmed `tokenize()`
