+++ assistant/src/plugins/defaults/memory/v2/sparse-bm25.ts
@@ -31,7 +31,7 @@ import {
   SPARSE_VOCAB_SIZE,
   tokenHash,
   tokenizeStemmed,
-} from "../sparse-tokenize.js";
+} from "../../../../persistence/embeddings/sparse-tokenize.js";
 import { listPages } from "./page-store.js";
 
 /**
